requirements.txt
aiomysql==0.0.20
aiosqlite==0.15.0
astroid==2.4.2
autopep8==1.5.4
cffi==1.14.2
click==7.1.2
cryptography==3.1
fastapi==0.61.1
h11==0.9.0
httptools==0.1.1
iso8601==0.1.12
isort==5.4.2
lazy-object-proxy==1.4.3
loguru==0.5.1
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
pylint==2.6.0
PyMySQL==0.9.2
PyPika==0.38.0
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.8
toml==0.10.1
tortoise-orm==0.16.14
typed-ast==1.4.1
typing-extensions==3.7.4.3
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
wrapt==1.12.1
